(fix) Trivial refactor in relationship component (#1590)
